refactor: generic toast notification [#135]
