+++ src/lib/components/Modal/CollectionPopup.svelte
@@ -7,7 +7,7 @@
   import { CollectionService } from "$lib/services/collection.service";
   import { ItemType } from "$lib/utils/enums/item-type.enum";
   import type { CollectionsMethods } from "$lib/utils/interfaces/collections.interface";
-  import { notifications } from "$lib/utils/notifications";
+  import { notifications } from "$lib/components/toast-notification/ToastNotification";
   import Button from "../buttons/Button.svelte";
   import { fly, fade } from "svelte/transition";
   export let collectionId: string;
