+++ src/lib/components/Modal/FolderPopup.svelte
@@ -6,7 +6,7 @@
   import closeIcon from "$lib/assets/close.svg";
   import { CollectionService } from "$lib/services/collection.service";
   import type { CollectionsMethods } from "$lib/utils/interfaces/collections.interface";
-  import { notifications } from "$lib/utils/notifications";
+  import { notifications } from "$lib/components/toast-notification/ToastNotification";
   import { fade, fly } from "svelte/transition";
   import Button from "../buttons/Button.svelte";
   export let collectionId: string;
