+++ src/lib/components/collections/collections-list/request/Request.svelte
@@ -15,7 +15,7 @@
   import { isApiCreatedFirstTime } from "$lib/store/request-response-section";
   import { setBodyType } from "$lib/utils/helpers/auth.helper";
   import ModalWrapperV1 from "$lib/components/Modal/Modal.svelte";
-  import { notifications } from "$lib/utils/notifications";
+  import { notifications } from "$lib/components/toast-notification/ToastNotification";
   import Button from "$lib/components/buttons/Button.svelte";
   import RightOption from "$lib/components/right-click-menu/RightClickMenuView.svelte";
