+++ src/lib/components/collections/req-res-section/SidebarRight.svelte
@@ -14,7 +14,7 @@
   import SaveIcon from "$lib/assets/save-desc.svg";
   import EditIcon from "$lib/assets/edit-desc.svg";
   import type { WorkspaceRole } from "$lib/utils/enums";
-  import { notifications } from "$lib/utils/notifications";
+  import { notifications } from "$lib/components/toast-notification/ToastNotification";
   import { workspaceLevelPermissions } from "$lib/utils/constants/permissions.constant";
   import { hasWorkpaceLevelPermission } from "$lib/utils/helpers";
   import ModalWrapperV1 from "$lib/components/Modal/Modal.svelte";
