+++ src/lib/components/dropdown/HeaderDropdown.svelte
@@ -12,7 +12,7 @@
     isWorkspaceLoaded,
   } from "$lib/store/workspace.store";
   import { UntrackedItems } from "$lib/utils/enums/item-type.enum";
-  import { notifications } from "$lib/utils/notifications";
+  import { notifications } from "$lib/components/toast-notification/ToastNotification";
   import { slide } from "svelte/transition";
   import { TickIcon } from "$lib/assets/app.asset";
   import type {
