+++ src/lib/components/environments/environments-list/EnvironmentList.svelte
@@ -8,7 +8,7 @@
   } from "$lib/utils/interfaces/environment.interface";
   import { onDestroy } from "svelte";
   import { UntrackedItems } from "$lib/utils/enums/item-type.enum";
-  import { notifications } from "$lib/utils/notifications";
+  import { notifications } from "$lib/components/toast-notification/ToastNotification";
   import { isEnvironmentCreatedFirstTime } from "$lib/store/environment";
   import Spinner from "$lib/components/Transition/Spinner.svelte";
   import { isWorkspaceLoaded } from "$lib/store/workspace.store";
