+++ src/lib/components/workspace/invite-to-workspace/InviteToWorkspace.svelte
@@ -9,7 +9,7 @@
     addUsersInWorkspace,
     addUsersInWorkspacePayload,
   } from "$lib/utils/dto";
-  import { notifications } from "$lib/utils/notifications";
+  import { notifications } from "$lib/components/toast-notification/ToastNotification";
   import { WorkspaceRole } from "$lib/utils/enums";
   import { createDynamicComponents } from "$lib/utils/helpers/common.helper";
   const emailstoBeSentArr: string[] = [];
