+++ src/pages/Home/Waiting.svelte
@@ -3,7 +3,7 @@
   import { setUser } from "$lib/store/auth.store";
   import constants from "$lib/utils/constants";
   import { jwtDecode, setAuthJwt } from "$lib/utils/jwt";
-  import { notifications } from "$lib/utils/notifications";
+  import { notifications } from "$lib/components/toast-notification/ToastNotification";
   import { onMount } from "svelte";
   import { navigate } from "svelte-navigator";
